+++ spring-framework-final-project/src/main/java/com/example/spring_framework_final_project/model/hotel/HotelResponse.java
@@ -13,7 +13,7 @@ public class HotelResponse {
     private String title;
     private String city;
     private String address;
-    private Integer distanceFromCentre;
+    private Double distanceFromCentre;
     private Double rating;
     private Long numberOfRatings;
 }
